Deps: Use bam 4.0 and drop explicit dependency to bam-ppx
* What

Update our dependency tree to use bam 4.0.

* Why

Bam 4.0 as released in the opam repository restricts the dependency of tezt-bam
to bam-ppx only for tests.

This is of interest for us because bam-ppx has a dependency to a ppxlib version
that we are not compatible with.

* How

Update the Opam repository hash in `version.sh` and run
`scripts/update_opam_lock.sh`. We only pick the changes related to bam here.