Deps: Bump ocamlformat to 0.27.0
* What

This commit upgrades the `ocamlformat` dependency from version 0.26.2
to 0.27.0. The change is applied across the project's dependency
management files and formatter configurations.

* Why

Version 0.27.0 is sadly a requirement of OCaml 5.3, as the version we are
currently using (0.26.2) is [not compatible with this release of
OCaml][commit].

[commit]: https://github.com/ocaml/opam-repository/commit/84e7d65d20130937e03acd5ef7ef76542da5bb10

* How

The `ocamlformat` version has been updated to `0.27.0` in the main
dependency manifest (`manifest/externals.ml`).

All `.ocamlformat` configuration files have been updated to reflect
this new version. Additionally, the option `parse-docstrings=false` has
been added to these configuration files to maintain the existing
formatting of docstrings.