Deps: Bump dream to a patched 1.0.0~alpha8
* What

This commit upgrades the `dream` web framework dependency from version
`1.0.0~alpha7` to a patched alternative of `1.0.0~alpha8`. As a
consequencer, this change also involves pinning `dream` to a specific
commit from an ad-hoc repository.

* Why

`dream.1.0.0~alpha7` is not compatible with OCaml 5.3, but the
latest release supposedly is. Unfortunately, `dream.1.0.0~alpha8`
cannot be installed due to incompatible dependency requirements
in `dream-httpaf.1.0.0~alpha4` (which both requires **and** is
incompatible with `h2.0.13.0`.

To fix this issue, we pin the `dream` library to a specific commit,
which addresses the `h2` issue. This proactive update (without
waiting for the release of a `alpha9` release) unblocks the migration
to OCaml 5.3.

* How

The `dream` version constraint was updated across all relevant `opam`
files. To use a specific unreleased version, `dream` and its companion
package `dream-httpaf` are now explicitly pinned to a commit on a
development fork within the opam lock update script. New virtual opam
lock files for these pinned packages have been created and integrated
into the Docker-based build and CI pipelines. Consequently, the main
dependency lock file, `octez-deps.opam.locked`, has been regenerated.

To make it explicit that we are not using the upstream release,
we append `-octez` to the official release names
(`dream.1.0.0~alpha8-octez` and `dream-httpaf.1.0.0~alpha4-octez`).

Co-authored-by: Mattias Roux <mattias.roux@nomadic-labs.com>