MIR/Typechecker/Tests: always use the parser
In the typechecker tests, we sometimes use the parser and sometimes
use the macros from micheline::test_helpers to build values of type
Micheline<'_>.

I don't totally understand why but generalizing `Ctx` to `impl _` in
the typechecker breaks the lifetime inference of the Micheline nodes
built by the macros.

This commit rewrites all typechecker tests to use the parser instead
of the macros.