CI: add [schedule_merge_train_cache_refresh] pipeline
Adds a scheduled pipeline that for now is a clone of a [merge_train] pipeline.

One difference that is already added is that this scheduled pipeline does not stop after a job failure or if new commits are added.

The goal of this pipeline is to be able to reset caches used in merge-train pipelines to avoid OOMs and speed-up jobs.

In future commits, we should make the following changes to this pipeline:
- the policy of job-specific caches will change from `pull-push` to `push-only`
- no changesets will be used so that jobs are always triggered (NB: probably not needed in scheduled pipelines but makes the pipeline clearer).