CI: add [trigger] dependency to [oc.build_kernels]
Surprisingly, adding a need to trigger adds in the needs both the trigger and the docker ci image job.

We should probably modify CIAO so that if you depend on a job, it is added in the needs as well.
NB: I suspect the pb only happens when the dependency is a Docker CI image job. It probably is due to a bug in the generation of these CI image jobs