EVM Node: Remove wasmer-c-api dependency
* What

This change completely removes the `wasmer-c-api` dependency from the
`wasm_runtime` crate. This simplifies the crate's dependency graph
and, as a consequence, slims down its public API by no longer
re-exporting symbols that are not needed.

* Why

Re-exporting the `wasmer_c_api` crate was a requirement inherited from
the Rollup Node Fast Exec. Even if we did not use the Fast Exec,
it was linked to the EVM node and therefore required these symbols to
be provided. Since then, we have reworked the dependency tree and this
requirement is no longer true.

* How

The dependency is removed by deleting its declaration from the
`Cargo.toml` file and removing the corresponding `pub use` statement
from `lib.rs`. This change also cleans up numerous transitive
dependencies from the project, as reflected in the updated `Cargo.lock`
file.