EVM Kernel: Decouple KernelHost from Wasm runtime implementation
* What

The `KernelHost` is modified to no longer be hardcoded with a specific
Wasm runtime implementation for internal host functions. Instead, the
`InternalRuntime` trait implementation is now injected as a dependency
during the `KernelHost`'s initialization.

* Why

This change primarily enhances the process of enabling native execution
for a kernel. By decoupling the core kernel logic from the concrete
host environment, we can now substitute the implementation of the
`InternalRuntime` provided by the WASM PVM by the one provided by the
EVM node. This will make preparing the native execution for Etherlink
5.0 a lot easier.

* How

The `InternalHost` struct, which provides the Wasm implementation of
the internal runtime, is renamed to `WasmInternalHost` for clarity.
The `KernelHost::init` function is made generic and updated to accept an
instance of an `InternalRuntime` implementer. Consequently, the main
kernel entrypoints, `kernel_loop` and `populate_delayed_inbox`, are
refactored to inject the `WasmInternalHost` by default, while exposing
new generic functions that can be used directly in tests with a mock
runtime.