EVM Node: Fix transactions getting stucked on sequencer downtime
* What

This commit modifies the transaction queue's processing loop to ensure
that expired transactions are removed from the pending set *before* a
new batch is sent for injection.

* Why

When the sequencer has a downtime, it is not uncommon for partners
sending many transactions to report “stucked” transactions, that
is transactions returned by `eth_getTransactionByHash` forever.

This is because the queue of the tx queue worker is modified
*before* trying to send transactions, but the transaction objects
table is cleared *after* this call. As a consequence, if the
HTTP request to inject transactions fails, the transaction objects
table is never cleared.

* How

The change consists of reordering the operations within the main
processing loop in `tx_queue.ml`. The call to
`Pending_transactions.drop` and the subsequent notification to
dropped transaction callbacks are moved to execute before the
`send_transactions_batch` function is invoked. This ensures
a failing sequencer don’t put the tx queue in a bad state.