Tezlink/Kernel: Extract default context macro from the test
This macro is useful to create a default context. For Tezlink, it
will be useful for the origination as we don't have have/need a full
context. THe only part of the context that need to be correct are the host
to store keys and the context::Context to have correct paths