EVM Node: Make transaction injection asynchronous
* What

This commit refactors the transaction queue to handle the injection
of transaction batches asynchronously. The worker no longer blocks
while waiting for the EVM node endpoint to confirm receipt of the
transactions, making the process non-blocking.

* Why

Previously, the `Tick` handler, which is responsible for selecting
and injecting transactions, would make a blocking RPC call to the EVM
node. If the endpoint was slow or unresponsive, the entire tx-queue
worker would stall, unable to process other incoming requests. Making
the injection process non-blocking improves the overall responsiveness
and throughput of the transaction queue.

* How

The `send_transactions_batch` function is modified to run within an
`Lwt.dont_wait` block in the `Tick` handler. A new map,
`waiting_injection`, is added to the worker's state to keep track of
transactions that have been sent but are awaiting confirmation.

Upon receiving a response from the EVM node, `send_transactions_batch`
now pushes a new internal `Injection_confirmation` request back into
the worker's own queue. A new handler for this request looks up the
original transaction in the `waiting_injection` map, executes its
callback, and removes it from the map, completing the asynchronous
flow.