EVM Node: Add dedicated Docker image and build script
* What

This commit introduces a dedicated, release-ready Docker image for the
`octez-evm-node`. It also adds a new script to automate the process of
building and publishing this image for multiple architectures.

* Why

This new approach has three main advantages over the current scripts
run on new Etherlink release:

  - Move to a dedicated repository `tezos/octez-evm-node` (instead of
    reusing existing ones normally dedicated to Octez)
  - Standalone process for releasing Docker images for the latest
    EVM release allows to update the base image without the need
    to re-release the binary
  - Trivial addition of new Alpine packages to the image without tampering with
    Octez releases

* How

The implementation consists of two new files. A `Dockerfile.release` is
added to define the image structure. It uses an Alpine base, copies the
`octez-evm-node` static binary, and installs a few essential
utilities. A new shell script, `build_docker_release.sh`, automates
the build. It fetches the latest `octez-evm-node` binary from GitLab
releases, verifies its checksum, and then executes `docker build` to
create and tag the final image for both `x86_64` and `arm64`
architectures.