Proto/DAL: do not record DAL content for the first levels of the protocol
Concretely, for the first `attestation_lag` levels of the protocol.

NOTE: this is only meant to be done in U. Since the value of this parameter
changes in U, this avoid corner cases related to the fact that the same attested
level has two corresponding publication levels, dependong on which
`attestation_lag` value is used.