EVM Node: Update the WASM Runtime to make use of Ebisu
This commit is larger than expected because being able to implement the
`InternalRuntime` trait of Ebisu along the `Runtime` trait of the SDK
requires the two implementors to share a mutable instance of the tree.

This is done once and for all by wrapping the tree owned by the `Host` inside a
`Rc<RefCell<_>>` wrapper that can be cloned to be given by the so-called
`Hasher` (the implementor of `InternalRuntime`).