Ebisu: Pin the MIR dependency to a specific commit
The commit picked for pinning the dependency is not the one proposed to the
governance, but a patched revision changing how the Kernel SDK dependencies are
fetched (c6a2505). This is a necessary change because previous kernel have been
pinned to a specific revision of these dependencies that we want to keep.