L2Node/Tezlink: remove Operation.decode
Operation.decode does more than the deserializing, it does a bit of
validation (but not all of it) to get the other values necessary for
insertion (namely, length of batch, source and first counter). Those
values need to be checked by the validation anyway.

So rather than reading the operation several time (during decoding and
validation) and risking manipulating operation not fully checked, we
fold the deserialization into the validation. That we don't have a easy
way to get a value of type Operation.t without prevalidating it.