Dal_node: Node_context: Move drop attested levels at migration from RPC_server
Done for the symmetry with the published level at migration case, since we should keep both functions in the same place. Note however that it is not trivial to use only one of them, because the condition changes depending on which perspective you see the problem from (i.e. published level vs. attested level case).