Tezlink/Kernel: zip operations in validate_and_apply_operation
The zip_operations function, which combines applied operations with
their receipts, forces us to copy the operations: one copy is used to
compute the receipts and the other as input to zip_operations. To
avoid this copy, we want to change the interface of the functions
producing receipts so that they also return the operation.

This commit is the first step in this direction; it moves the call to
zip_operations from the kernel crate to the tezos_execution crate.