CIAO: turn [Code_verification.Opam] [pipeline_type] agnostic
For a while now, opam jobs have been removed from before_merging/merge_train
pipelines.

In !19789, a daily pipeline has been created for opam packages.

It makes sense now to make [Code_verification.Opam] pipeline-type agnostic as
it is non-necessary and can lead to code simplifications:
- no more pipeline_type in [job_opam_packages] and [jobs_opam_packages]
- no opam_rules
- simpler registration of [opam.daily]

NB: it removes the effect of applying the label `ci--opam` but in practice
it shouldn't change much as it was only used in [before_merging/merge_train] pipelines.