Rollups/Dal: Adapt echo/tx kernel and Arith to add some slack before fetching pages
With this MR, the rollup node relies on the DAL node to fetch slots statuses and
data. Since those data on the DAL node are stored once blocks are final, kernels
should wait 2 levels before requesting pages. One extra level is needed to let
the DAL node index the data.