Rollup Node: Apply specific DAL activation level for Shadownet
* What

This commit introduces a special case for the "Shadownet" test
network. It hardcodes the DAL activation level to 1 when the
rollup node is connected to this specific chain. This logic
is applied within the DAL slot and page validation functions.

* Why

Shadownet is a test environment where the DAL is misconfigured to
activate at level 5,726,209. To ensure Shadownet remains a
place where we can test deployment relaying on the DAL, this patch
overrides the configured `dal_activation_level` when the Shadownet
chain ID is detected.

Obviously, this only “works” because no real Tez are at stake on
Shadownet.

* How

A new constant, `shadownet`, is defined in
`dal_pages_request.ml` using its B58-encoded chain ID.

The `slot_id_is_valid` function is modified to accept a
`chain_id` parameter. Inside this function, it checks if the
provided ID matches the `shadownet` constant. If it does, it
locally overrides the `dal_activation_level` to be level 1.

The `chain_id` is now retrieved from the `l1_ctxt` in
`slot_pages` and `page_content` using `Layer1.get_chain_id`.
This ID is then passed down to `page_id_is_valid` and
`slot_id_is_valid`, ensuring the correct validation logic is
applied based on the connected L1 chain.

This patch is applied for both Seoul and Alpha.