Lib/Wasm: use macro for reveal metadata so that it can be disabled via the parameter
At the time I forgot to wrap the function so I can use the macro from the
library to trace it.
Because of that if I disabled the host func tracing with otel, it escapes
the parameter since it wasn't trace via the macro.

Also fixed the typo: reaveal -> reveal.