Tezlink/L2Node: tweak simulation mock to return valid gas limit
The mock now define a safe value for the gas limit that is the hard cap
minus 1000 gas, to account for clients adding a bit more. For batch,
that value is distributed among all operations in the batch, to ensure
the batch is under the limit.
This is very crude, but also a mock.