EVM Kernel: Prevent event generation when run by the EVM node
* What

This change modifies the kernel runtime to cache the EVM node
configuration flag in memory and uses it to suppress the persistence of
events when the kernel is executed by the EVM node.

* Why

Events are intended as a communication mechanism from the rollup node to
the EVM node. When the EVM node is the one executing the kernel, writing
these events to storage is redundant and creates unnecessary I/O
overhead. Furthermore, events garbage collection is performed during
the stage one, but the EVM node does not run this stage anymore, meaning
the storage actually accumulates useless data in the context.

This new behavior is advertised by the introduction of the 43th storage
version.

* How

The patch introduces a new `IsEvmNode` trait in the runtime and expands
the `KernelHost` struct to include an `is_evm_node` boolean. This field
is initialized in `KernelHost::init` by reading the `/__evm_node` flag
from storage a single time. The `Event::store` method is updated to
check this cached value and skip `storage::store_event` if the kernel is
running within the EVM node. Finally, previous direct calls to
`evm_node_flag` in `block.rs` and `configuration.rs` are replaced with
the new trait method.