Kernel/SDK: restore `tezos-smart-rollup::types::Contract` alias
!19958 removed the `Contract` type from the KernelSDK as
`tezos-protocol` contained a duplicated implementation.

Removing this was a backwards-incompatible change, however - and
therefore needs to be recorded in the changelog. Additionally, the
top-level alias in `tezos-smart-rollup` needs to be restored, as the
top-level crate is designed as a (mostly) one stop shop for kernels
needs. In particular, this change breaks jstz when upgrading.