Dockerfile: upgrade system packages systematically
We can usually get the latest versions of the packages, e.g. forcing a fresh build with no-cache.

For system packages though, we will only get the version installed in the Alpine image.

Newer versions may be available but not yet included in the Alpine image.

It is actually currently the case now for [busybox]:
- alpine:3.20@sha256:765942a4039992336de8dd5db680586e1a206607dd06170ff0a37267a9e01958
- includes busybox-1.36.1-r30 which has vulns (CVE-2024-58251, CVE-2025-46394)
- but busybox-1.36.1-r31 is available and fixes the vulns.